feat: add SpyCredentialsStorage for testing credential storage behavior
